# 003 Code/Aiary-Backend-Public-main/src/main/java/org/example/aiarycproject/domain/user/dto/LoginRequest.java
package org.example.aiarycproject.domain.user.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class LoginRequest {

    @JsonProperty("user_name")
    private String userName;

    private String password;
}