# 003 Code/Aiary-Backend-Public-main/swagger-api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/chat/threads": {
      "get": {
        "tags": [
          "Chat Thread"
        ],
        "summary": "스레드 목록 조회",
        "description": "상태별 스레드 목록을 페이지네이션으로 조회합니다",
        "operationId": "list_1",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "스레드 상태 (active/archived)",
            "required": false,
            "schema": {
              "type": "string",
              "default": "active"
            },
            "example": "active"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "조회할 항목 수 (최대 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            },
            "example": 20
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "페이지네이션 커서 (이전 응답의 nextCursor 값)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "스레드 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Chat Thread"
        ],
        "summary": "스레드 생성",
        "description": "새로운 채팅 스레드를 생성합니다",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateThreadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "스레드 생성 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/sign-up": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "signUp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/chat/threads/{threadId}": {
      "get": {
        "tags": [
          "Chat Thread"
        ],
        "summary": "스레드 조회",
        "description": "특정 스레드의 상세 정보를 조회합니다",
        "operationId": "get_1",
        "parameters": [
          {
            "name": "threadId",
            "in": "path",
            "description": "스레드 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "스레드를 찾을 수 없음",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatThread"
                }
              }
            }
          },
          "200": {
            "description": "스레드 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatThread"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Chat Thread"
        ],
        "summary": "스레드 삭제",
        "description": "특정 스레드를 삭제합니다",
        "operationId": "delete",
        "parameters": [
          {
            "name": "threadId",
            "in": "path",
            "description": "스레드 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "스레드를 찾을 수 없음",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "204": {
            "description": "스레드 삭제 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Chat Thread"
        ],
        "summary": "스레드 상태 변경",
        "description": "스레드의 상태를 변경합니다 (active/archived)",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "threadId",
            "in": "path",
            "description": "스레드 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "상태 변경 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "스레드를 찾을 수 없음",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateThreadRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          }
        }
      },
...
      "SignUpRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "user_name": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "user_name": {
            "type": "string"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nickname": {
            "type": "string"
          },
          "user_name": {
            "type": "string"
          }
        }
      },
...
      "UpdateStatusRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "ChatThread": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}